refactor towards incremental plans
