+++ nexus/deployment/Cargo.toml
@@ -11,6 +11,7 @@ ipnet.workspace = true
 ipnetwork.workspace = true
 nexus-types.workspace = true
 omicron-common.workspace = true
+slog.workspace = true
 thiserror.workspace = true
 uuid.workspace = true
